requirements.txt
langgraph
langchain 
langchain-groq 
python-dotenv
langchain-chroma
langchain-huggingface
sentence-transformers